requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
attrs==21.2.0
av==8.0.3
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
clip==1.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
ftfy==6.0.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.6.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
regex==2021.8.21
requests==2.26.0
scikit-video==1.1.11
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
streamlit==0.87.0
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
